requirements.txt
python-telegram-bot==21.7
Telethon==1.36.0
python-dotenv==1.0.1
